+++ packages/connection-encrypter-tls/CHANGELOG.md
@@ -6,6 +6,24 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [2.2.6](https://github.com/libp2p/js-libp2p/compare/tls-v2.2.5...tls-v2.2.6) (2025-06-25)
+
+
+### Bug Fixes
+
+* check for signal abort in tls during cert generation ([#3203](https://github.com/libp2p/js-libp2p/issues/3203)) ([82ac83c](https://github.com/libp2p/js-libp2p/commit/82ac83c0d532abf95cc17debea7e7b208ee0a8aa))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+
 ## [2.2.5](https://github.com/libp2p/js-libp2p/compare/tls-v2.2.4...tls-v2.2.5) (2025-06-16)