+++ packages/metrics-prometheus/CHANGELOG.md
@@ -81,6 +81,20 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [4.3.26](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.25...prometheus-metrics-v4.3.26) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface-compliance-tests bumped from ^6.4.12 to ^6.4.13
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+
 ## [4.3.25](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.24...prometheus-metrics-v4.3.25) (2025-06-18)