+++ packages/peer-record/CHANGELOG.md
@@ -60,6 +60,18 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [8.0.34](https://github.com/libp2p/js-libp2p/compare/peer-record-v8.0.33...peer-record-v8.0.34) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+    * @libp2p/utils bumped from ^6.7.0 to ^6.7.1
+
 ## [8.0.33](https://github.com/libp2p/js-libp2p/compare/peer-record-v8.0.32...peer-record-v8.0.33) (2025-06-17)