+++ packages/pnet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [2.0.43](https://github.com/libp2p/js-libp2p/compare/pnet-v2.0.42...pnet-v2.0.43) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.12 to ^6.4.13
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+
 ## [2.0.42](https://github.com/libp2p/js-libp2p/compare/pnet-v2.0.41...pnet-v2.0.42) (2025-06-18)