+++ packages/protocol-fetch/CHANGELOG.md
@@ -48,6 +48,20 @@
   * devDependencies
     * @libp2p/peer-id-factory bumped from ^4.0.3 to ^4.0.4
 
+## [3.0.21](https://github.com/libp2p/js-libp2p/compare/fetch-v3.0.20...fetch-v3.0.21) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/interface-internal bumped from ^2.3.17 to ^2.3.18
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+
 ## [3.0.20](https://github.com/libp2p/js-libp2p/compare/fetch-v3.0.19...fetch-v3.0.20) (2025-06-17)